+++ internal/Vk/TextureAtlasVK.cpp
@@ -388,7 +388,7 @@ bool Ray::Vk::TextureAtlas::Resize(const int pages_count) {
         VkCommandBuffer cmd_buf = BegSingleTimeCommands(ctx_->api(), ctx_->device(), ctx_->temp_command_pool());
 
         ctx_->api().vkCmdPipelineBarrier(cmd_buf, src_stages ? src_stages : VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                                         dst_stages, 0, 0, nullptr, 0, nullptr, uint32_t(img_barriers.size()),
+                                         dst_stages, 0, 0, nullptr, 0, nullptr, img_barriers.size(),
                                          img_barriers.cdata());
 
         resource_state = eResState::CopySrc;
@@ -454,10 +454,10 @@ bool Ray::Vk::TextureAtlas::Resize(const int pages_count) {
 
         VkCommandBuffer cmd_buf = BegSingleTimeCommands(ctx_->api(), ctx_->device(), ctx_->temp_command_pool());
 
-        ctx_->api().vkCmdPipelineBarrier(
-            cmd_buf, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-            VKPipelineStagesForState(eResState::ShaderResource) & ctx_->supported_stages_mask(), 0, 0, nullptr, 0,
-            nullptr, uint32_t(img_barriers.size()), img_barriers.cdata());
+        ctx_->api().vkCmdPipelineBarrier(cmd_buf, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
+                                         VKPipelineStagesForState(eResState::ShaderResource) &
+                                             ctx_->supported_stages_mask(),
+                                         0, 0, nullptr, 0, nullptr, img_barriers.size(), img_barriers.cdata());
 
         EndSingleTimeCommands(ctx_->api(), ctx_->device(), ctx_->graphics_queue(), cmd_buf, ctx_->temp_command_pool());
     }
@@ -514,7 +514,7 @@ int Ray::Vk::TextureAtlas::DownsampleRegion(const int src_page, const int src_po
         dst_stages &= ctx_->supported_stages_mask();
 
         ctx_->api().vkCmdPipelineBarrier(cmd_buf, src_stages ? src_stages : VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                                         dst_stages, 0, 0, nullptr, 0, nullptr, uint32_t(img_barriers.size()),
+                                         dst_stages, 0, 0, nullptr, 0, nullptr, img_barriers.size(),
                                          img_barriers.cdata());
     }
 
@@ -568,7 +568,7 @@ int Ray::Vk::TextureAtlas::DownsampleRegion(const int src_page, const int src_po
         dst_stages &= ctx_->supported_stages_mask();
 
         ctx_->api().vkCmdPipelineBarrier(cmd_buf, src_stages ? src_stages : VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                                         dst_stages, 0, 0, nullptr, 0, nullptr, uint32_t(img_barriers.size()),
+                                         dst_stages, 0, 0, nullptr, 0, nullptr, img_barriers.size(),
                                          img_barriers.cdata());
     }
 
@@ -742,7 +742,7 @@ int Ray::Vk::TextureAtlas::DownsampleRegion(const int src_page, const int src_po
         dst_stages &= ctx_->supported_stages_mask();
 
         ctx_->api().vkCmdPipelineBarrier(cmd_buf, src_stages ? src_stages : VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                                         dst_stages, 0, 0, nullptr, 0, nullptr, uint32_t(img_barriers.size()),
+                                         dst_stages, 0, 0, nullptr, 0, nullptr, img_barriers.size(),
                                          img_barriers.cdata());
     }
 
@@ -906,8 +906,8 @@ void Ray::Vk::TextureAtlas::CopyRegionTo(const int page, const int x, const int 
 
     if (!buf_barriers.empty() || !img_barriers.empty()) {
         ctx_->api().vkCmdPipelineBarrier(cmd_buf, src_stages ? src_stages : VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                                         dst_stages, 0, 0, nullptr, uint32_t(buf_barriers.size()), buf_barriers.cdata(),
-                                         uint32_t(img_barriers.size()), img_barriers.cdata());
+                                         dst_stages, 0, 0, nullptr, buf_barriers.size(), buf_barriers.cdata(),
+                                         img_barriers.size(), img_barriers.cdata());
     }
 
     resource_state = eResState::CopySrc;
